requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
sentence-transformers
pdfminer.six
python-docx
python-multipart
aiofiles

